+++ lw10/Lw10/lib/SimpleShapeFactory/SimpleShapes.h
@@ -13,9 +13,9 @@ protected:
 	void DrawImp(Canvas::ICanvas& canvas) const override
 	{
 		Point p1 = { GetRect().left, GetRect().top };
-		Point p2 = { GetRect().left, GetRect().top + GetRect().height };
-		Point p3 = { GetRect().left + GetRect().width, GetRect().top + GetRect().height };
-		Point p4 = { GetRect().left + GetRect().width, GetRect().top };
+		Point p2 = { GetRect().left, GetRect().top + static_cast<int>(GetRect().height) };
+		Point p3 = { GetRect().left + static_cast<int>(GetRect().width), GetRect().top + static_cast<int>(GetRect().height) };
+		Point p4 = { GetRect().left + static_cast<int>(GetRect().width), GetRect().top };
 
 		canvas.DrawPolygon({ p1, p2, p3, p4 });
 	}
@@ -35,9 +35,9 @@ class Triangle : public Shape
 protected:
 	void DrawImp(Canvas::ICanvas& canvas) const override
 	{
-		Point p1 = { GetRect().left + GetRect().width / 2, GetRect().top };
-		Point p2 = { GetRect().left, GetRect().top + GetRect().height };
-		Point p3 = { GetRect().left + GetRect().width, GetRect().top + GetRect().height };
+		Point p1 = { GetRect().left + static_cast<int>(GetRect().width) / 2, GetRect().top };
+		Point p2 = { GetRect().left, GetRect().top + static_cast<int>(GetRect().height) };
+		Point p3 = { GetRect().left + static_cast<int>(GetRect().width), GetRect().top + static_cast<int>(GetRect().height) };
 
 		canvas.DrawPolygon({ p1, p2, p3 });
 	}
